Restructured run.py to individual classes
